Add files via upload
The Electric Vehicles Data Analysis Dashboard, created using Tableau, provides a clear and insightful overview of the electric vehicle (EV) landscape in the United States. With a total of 150,413 EVs analyzed, the dashboard highlights key statistics such as an average electric range of 67.83 miles and a breakdown of vehicle types, showing that 77.6% are Battery Electric Vehicles (BEVs) and 22.4% are Plug-in Hybrid Electric Vehicles (PHEVs). Tesla stands out as the leading manufacturer, capturing 55.6% of the market with 68,939 vehicles, followed by Nissan and Chevrolet. Geographically, California emerges as the EV hotspot, boasting a substantial 150,082 vehicles. Popular models like the Tesla Model Y and Model 3 lead the charts with 28,501 and 27,708 vehicles sold, respectively. The dashboard’s visualizations make it easy to understand market trends, regional adoption rates, and the most preferred EV models. This tool is designed to help analysts, businesses, and policymakers make well-informed, data-driven decisions by offering a clear snapshot of the current EV market dynamics.